Switch to makefile for building tfm

+++ tools/env_setup.sh
@@ -30,15 +30,24 @@ if [ -n "${BASH_SOURCE[0]}" ]; then
 else
 	TOOLS_PATH=$( realpath "$(cd "$(dirname "${0}")" || return; pwd )" )
 fi
-echo "TOOLS_PATH: ${TOOLS_PATH}"
+
+(return 0 2>/dev/null) || {
+	echo "WARNING: To work properly, this script should be sourced rather than run as a command."
+	DEBUG=1
+}
+
+[ -n "${DEBUG}" ] && echo "TOOLS_PATH:         ${TOOLS_PATH}"
 
 VENV_PATH=$(realpath "${TOOLS_PATH}/..")/.venv
-echo "VENV_PATH:  ${VENV_PATH}"
-echo
+[ -n "${DEBUG}" ] && echo "VENV_PATH:          ${VENV_PATH}"
+
+MUCBOOT_PATH=$(realpath "${TOOLS_PATH}/..")/Middleware/ARM/mcuboot
+[ -n "${DEBUG}" ] && echo "MCUBOOT_PATH:       ${MUCBOOT_PATH}"
+[ -n "${DEBUG}" ] && echo
 
 which python > /dev/null 2>&1 || alias python=python3
 which python > /dev/null 2>&1 || {
-	echo "Failed to find a validate python installation."
+	echo "Error: Failed to find a valid python installation."
 	return 1
 }
 
@@ -52,24 +61,47 @@ load_venv()
 		echo "Error. Could not find python virtual environment activation script."
 		return 1
 	fi
-
-	echo "Adding tools directory to path."
-	export PATH="${TOOLS_PATH}:${PATH}"
 }
 
 if [ ! -d "${VENV_PATH}" ]; then
 	echo "Setting up python virtual environment"
 	python -m virtualenv "${VENV_PATH}" || return 1
+fi
 
-	echo "Activating vritual environment."
-	load_venv || return 1
+[ -n "${DEBUG}" ] && echo "Activating vritual environment."
+load_venv || return 1
 
-	pip3 install -r "${TOOLS_PATH}"/requirements.txt || {
+# .initialized has not been written or requirements.txt / env_setup.sh have been modified
+if [[ ! -e "${VENV_PATH}"/.initialized ]] || \
+   [[ "${TOOLS_PATH}/requirements.txt" -nt "${VENV_PATH}/.initialized" ]] || \
+   [[ "${TOOLS_PATH}/env_setup.sh" -nt "${VENV_PATH}/.initialized" ]]; then
+	python -m pip install -r "${TOOLS_PATH}"/requirements.txt || {
 		rm -rf "${VENV_PATH}"
 		echo "Error while installing requried python packages. Removing incomplete virtual environment."
 		return 1
 	}
-else
-	load_venv || return 1
+
+	SITE_PACKAGES_PATH=$(find ${VENV_PATH} -name site-packages -type d)
+	[ -n "${DEBUG}" ] && echo "SITE_PACKAGES_PATH: ${SITE_PACKAGES_PATH}"
+
+	echo "Adding tools directory to package path."
+	[ -n "${DEBUG}" ] && echo ${TOOLS_PATH} > ${SITE_PACKAGES_PATH}/tools.pth
+
+	echo "Adding mcuboot scripts directory to package path."
+	[ -n "${DEBUG}" ] && echo ${MCUBOOT_PATH}/scripts > ${SITE_PACKAGES_PATH}/mcuboot.pth
+
+	touch "${VENV_PATH}"/.initialized
 fi
-echo "Done."
+
+SITE_PACKAGES_PATH=$(find "${VENV_PATH}" -name site-packages -type d)
+
+# Setup PATH and PYTHONPATH as a workaround for cmake (which doesn't appear to respect virtual environments fully)
+[ -n "${DEBUG}" ] && echo "Adding tools and site packages to PATH."
+export PATH="${TOOLS_PATH}:${SITE_PACKAGES_PATH}:${PATH}"
+
+[ -n "${DEBUG}" ] && echo "Adding tools and site packages to PYTHONPATH."
+export PYTHONPATH="${TOOLS_PATH}:${SITE_PACKAGES_PATH}:${PYTHONPATH}"
+
+[ -n "${DEBUG}" ] && echo "Done."
+
+[ -n "${DEBUG}" ] && exit 0 || return 0
